refactor code into modules
