+++ src/pg_rad/isotopes/isotope.py
