+++ src/pg_rad/objects/sources.py
@@ -1,7 +1,7 @@
 import logging
 
 from pg_rad.objects import Object
-from pg_rad.isotope import Isotope
+from pg_rad.isotopes import Isotope
 
 logger = logging.getLogger(__name__)
